Change type of duration
To remove some Warnings

+++ Effect.h
@@ -10,7 +10,7 @@ private :
 	sf::Color FlashScreenColor = sf::Color::Black;
 
 public:
-	void ChangeFlashScreen(int duration, bool randomness, sf::Color color = sf::Color::Black);
+	void ChangeFlashScreen(float duration, bool randomness, sf::Color color = sf::Color::Black);
 	sf::Color RandomColor();
 
 	void UpdateEffect(sf::RenderWindow& window, float deltatime);
